Change Player::getId() and MyPlayerId to uint8_t
(cherry picked from commit fae6b3d8178698aee2218e8da34cf5432d8bec2e)

+++ Source/player.h
@@ -422,7 +422,7 @@ struct Player {
 	/**
 	 * @brief Returns the network identifier for this player
 	 */
-	[[nodiscard]] size_t getId() const;
+	[[nodiscard]] uint8_t getId() const;
 
 	void RemoveSpdBarItem(int iv);
 
@@ -826,7 +826,7 @@ struct Player {
 	uint32_t GetGolemToHit() const;
 };
 
-extern DVL_API_FOR_TEST size_t MyPlayerId;
+extern DVL_API_FOR_TEST uint8_t MyPlayerId;
 extern DVL_API_FOR_TEST Player *MyPlayer;
 extern DVL_API_FOR_TEST std::vector<Player> Players;
 /** @brief What Player items and stats should be displayed? Normally this is identical to MyPlayer but can differ when /inspect was used. */
